add counter for device

+++ service-device-management/src/main/java/com/sitewhere/device/persistence/mongodb/MongoDevice.java
@@ -122,7 +122,7 @@ public class MongoDevice implements MongoConverter<IDevice> {
 		String hardwareId = (String) source.get(PROP_HARDWARE_ID);
 		Object itemChannelLink = source.get(PROP_ITEM_CHANNEL_LINK);
 		Object configurationGateway = source.get(PROP_CONFIGURATION_GATEWAY);
-		Integer counter = (Integer) source.get(PROP_COUNTER);
+		Long counter = (Long) source.get(PROP_COUNTER);
 
 	target.setDeviceTypeId(typeId);
 	target.setParentDeviceId(parentDeviceId);
@@ -131,7 +131,7 @@ public class MongoDevice implements MongoConverter<IDevice> {
 	target.setDeviceAssignmentId(assignmentId);
 	target.setGatewayId(gatewayId);
 	target.setHardwareId(hardwareId);
-	target.setCounter(counter != null ? Long.valueOf(counter) : 0);
+	target.setCounter(counter != null ? counter : 0);
 	try {
 		Map<String, List<String>> itemChannelLinkView = (Map<String, List<String>>) itemChannelLink;
 		target.setItemChannelLink(itemChannelLinkView);
